requirements.txt
Django==1.6.2
Jinja2>=2.6
MySQL-python>=1.2.4
South>=0.7.6
argparse>=1.2.1
django-appconf>=0.6
django-debug-toolbar>=0.9.4
django-imagekit>=2.0.4
ipython>=0.13.1
redis>=2.7.2
six>=1.3.0
wsgiref>=0.1.2
django-redis==3.7.1
